- Clicking a strign in the editor now highlights it in the storyexplorer - reworked the clicked nodes as a event
