+++ Helpers/ClickedNodeChangeArgs.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace HousePartyTranslator.Helpers
+{
+    public class ClickedNodeChangeArgs : EventArgs
+    {
+        public Node HighlightNode { get; }
+
+        public ClickedNodeTypes HighlightCase { get; }
+
+        public ClickedNodeChangeArgs(Node highlightNode, ClickedNodeTypes highlightCase)
+        {
+            HighlightNode = highlightNode;
+            HighlightCase = highlightCase;
+        }
+    }
+}
